hadapi: Re-work forbidden handling.
Rather than a dumb wait of an hour, we simply set a flag that's exposed
elsewhere.  Our background tasks can check that flag instead and inhibit
their operations themselves.

If a user action that requires API access, succeeds, the flag is
cleared and the background tasks resume.

The background tasks auto-retry after an hour.