traits: Use thread-local storage for database objects.
SQLAlchemy really gets confused if you get objects from different
database sessions mixed up, causing things to not be recorded when they
should.

To counteract this, use thread-local storage for the connection and its
database objects.  This necessitates that we have a way to create new
connections to the database when a new thread is spawned.