hadapi: Re-work user ID scan.
This is a work-around to the fact that the API throws error 500 when you
try to retrieve the "newest" users (in spite of the fact that those are
generally the users that are of interest from a spam-prevention point of
view).

This is a last attempt to try and make this work.  Next time I just shut
down the project and let Hackaday.io drown in its spam, I spend too many
hours a week trying to clean up this crap and my own projects are
suffering because of it.